refactor: split big script into modules
