+++ src/main.py
@@ -1,7 +1,13 @@
+#!/usr/bin/env python3
+
 import numpy as np
-
 import pandas as pd
+import matplotlib.pyplot as plt
+import networkx as nx
+from utils import draw_pgm
+from core import estimate_parameters
 
+# this is sample data copied from the tutorial paper
 data = np.array([
     [3.626, 2.811, 17.683, 12.626],
     [1.895, 3.571, 15.320, 10.459],
@@ -14,68 +20,10 @@ data = np.array([
     [4.205, -0.090, 7.916,  6.756],
     [2.344, 2.823, 14.594, 10.383]
 ])
+coulumns_names=['X1', 'X2', 'X3', 'X4']
+df = pd.DataFrame(data, columns=coulumns_names)
+map_nodes_to_indexes = {node: i for i, node in enumerate(coulumns_names)}
 
-def estimate_parameters(structure, data) -> dict:
-    n_obs, n_vars = data.shape
-    fit_results = {}
-    for ivar, var_name in enumerate(structure.keys()):
-        dependent_vars = structure.get(var_name, [])
-        if len(dependent_vars)==0:
-            # then var is independent
-            intercept = np.mean(data[:, ivar])
-            residuals = data[:, ivar] - intercept
-            #print("RESIDUALS: ", np.var(residuals))
-            #print("RESIDUALS SQUARED: ", np.var(residuals)**2)
-            #print("SQRT RESIDUALS VARIANCE: ", np.sqrt(np.var(residuals)))
-            fit_results[ivar] = {
-                "intercept": intercept,
-                "coefficients": np.zeros(n_vars),
-                "residual_variance": np.var(residuals, ddof=1)
-            }
-        else:
-            mask = [map_nodes_to_indexes[var] for var in dependent_vars]
-            X = data[:, mask] 
-            y = data[:, ivar]
-            reg = LinearRegression(fit_intercept=True).fit(X, y)
-            intercept = reg.intercept_
-            coefficients = reg.coef_
-            residuals = y - reg.predict(X)
-            #print("RESIDUALS: ", np.var(residuals))
-            #print("RESIDUALS SQUARED: ", np.var(residuals)**2)
-            #print("SQRT RESIDUALS VARIANCE: ", np.sqrt(np.var(residuals)))
-            fit_results[ivar] = {
-                "intercept": intercept,
-                "coefficients": coefficients,
-                "residual_variance": np.var(residuals, ddof = X.shape[1])
-            }
-    return fit_results
-
-from pgmpy.base import DAG
-import matplotlib.pyplot as plt
-import networkx as nx
-def draw_pgm(ax, model, pos: dict = None, title: str = None):
-    #Init network instance
-    g = nx.DiGraph()
-
-    #Add nodes and edges
-    g.add_nodes_from(model.nodes())
-    g.add_edges_from(model.edges())
-
-    if pos is None:
-        pos = nx.circular_layout(G)
-
-    nx.draw(g, pos,
-        with_labels=True,
-        node_size = 1000, node_color = "skyblue",
-        font_size = 10, font_weight = "bold",
-        arrowsize=30, ax=ax)
-    ax.set_title(title)
-
-df = pd.DataFrame(data, columns=['X1', 'X2', 'X3', 'X4'])
-
-map_nodes_to_indexes = {node: i for i, node in enumerate(df.columns)}
-
-from sklearn.linear_model import LinearRegression
 
 edges_example_1 = [("X1","X4"),("X2","X4"),("X4","X3")]
 edges_example_2 = [("X1","X2"),("X2","X4"),("X2","X3")]
@@ -88,19 +36,22 @@ fit_results_list = []
 for istruct, edges_example in enumerate(structure_pair):
     print("STRUCTURE ", edges_example)
     G = nx.DiGraph()
-    G.add_nodes_from(['X1', 'X2', 'X3', 'X4'])
+    G.add_nodes_from(coulumns_names)
     G.add_edges_from(edges_example)
 
     fig, axs = plt.subplots(1, 1, figsize=(5, 5))
     draw_pgm(axs, G)
-    plt.show()
+    if plt.isinteractive():
+        plt.show()
+    else:
+        plt.savefig(f"image_{istruct}.png")
 
 ## INIT MODEL STRUCTURE BASED ON BAYESIAN NETWORK G STRUCTURE
     model_structure = {var: list(G.predecessors(var)) for var in G.nodes()}
     print("MODEL STRUCTURE: \n\t", model_structure)
 
 #### ESTIMATE PARAMETERS ####
-    fit_results = estimate_parameters(model_structure, data)
+    fit_results = estimate_parameters(model_structure, data, map_nodes_to_indexes)
     associated_data = np.zeros(data.shape)
     for ivar, var_name in enumerate(model_structure.keys()):
         dependent_vars = model_structure.get(var_name, [])
@@ -109,10 +60,10 @@ for istruct, edges_example in enumerate(structure_pair):
         else:
             associated_data[:,ivar] = fit_results[ivar]["intercept"] + np.dot(data[:, [map_nodes_to_indexes[var] for var in dependent_vars]], fit_results[ivar]["coefficients"])
 
-    #print("ASSOCIATED DATA: \n", associated_data)
+    # print("ASSOCIATED DATA: \n", associated_data)
     associated_data_list.append(associated_data)
     fit_results_list.append(fit_results)
-            
+
     #### COMPUTE SCORE ####
     bn_entropy = 0
     for var, results in fit_results.items():
